+++ src/main/java/jaid/collection/IVector.java
@@ -15,7 +15,7 @@ public interface IVector {
      * @param bits the number of bits to be used for bucket allocation
      * @return the bucket number the current vector is allocated to
      */
-    int simBucket(int bits);
+    int simBucket(byte bits);
 
     String toString();
 }
